backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "librosa>=0.11.0",
    "openai>=2.17.0",
    "pypdf>=6.6.2",
    "torch>=2.10.0",
    "transformers>=5.1.0",
]
